Pseudo-arclength continuation (PALC)

PALC

This contains information regarding pseudo-arclength continuation.

SimpleContinuation.PALC — Type

PALC

Data type for storing all information for PALC algorithm.

SimpleContinuation.PALC — Method

PALC(; kwargs...)

Contructor for PALC algorithm.

Kwargs

predicter::AbstractPredictor: Prediction method. Default (recommended): Bordered()

inner_product::AbstractInnerProduct: Inner product to use throughout. Default: StandardDotProduct()

linesearch: Line search method provided by LineSearch.jl. Default: LiFukushimaLineSearch()

linsolve: Linear solve algorithm provided by LinearSolve.jl. Default SVDFactorization()

termcond: Termination condition for solvers. Default: termcond=NonlinearSolve.AbsTerminationMode()

Examples

pred = Secant() # secant predictor
alg = PALC(; predicter=pred) # all defaults except prediction method

SimpleContinuation.PALCCache — Type

PALCCache{MT<:Union{Matrix{Float64},SparseMatrixCSC{Float64,Int}}}

A cache for the PALC algorithm.

This cache includes all preallocated storage required for PALC. It also contains the zero curve (solution), return code, and any special detected points. This cache is not directly exported, but it is returned by continuation methods.

Fields

ds::Float64: PALC step size

br::Vector{Tuple{Vector{Float64},Float64}}: Zero curve storage, in format [(u0,λ0); (u1,λ1); ...; (un,λn)] for n iterations.

detected_points::Vector{Tuple{Vector{Float64},Float64, Symbol}}: Storage for any points of interest detected by FoldBifurcationDetectionCallback() or any other detection callbacks (once implemented).

uλ0::Vector{Float64}: Full [u;λ] current iterate

u0::Vector{Float64}: Unknowns u current iterate.

λ0::Float64: Continuation parameter λ current iterate.

λn::Float64: Natural continuation parameter. Only used in select instances when required.

bordered_mat::MT: Bordered matrix storage

bordered_b::Vector{Float64}: Storage for RHS of bordered system.

δuλ0::Vector{Float64}: Storage for full tangent [δu; δλ]

δu0::Vector{Float64}: Storage for tangent components δu

δλ::Float64: Storage for tangent component δλ

δuλ0_initial::Vector{Float64}: Storage for initalized tangent computed before first iteration.

uλpred::Vector{Float64}: Storage for predicted point

Ffun::Vector{Float64}: Storage for continuation function residuals (not including hyperplane constraint)

Jfun::MT: Storage for non-square jacobian of system ∂F(u;λ)/∂[u;λ]. (not including hyperplane constraint)

u_0::Vector{Float64}: Storage for regula-falsi solver

u_1::Vector{Float64}: Storage for regula-falsi solver

u_t::Vector{Float64}: Storage for regula-falsi solver

ret::Symbol: Exit condition return code.

Predictors

SimpleContinuation.Bordered — Type

Bordered <: AbstractPredictor

Bordered prediction method. Uses the bordered matrix to compute predicted updates. This is the default for PALC.

SimpleContinuation.Secant — Type

Secant <: AbstractPredictor

Secant prediction method. Uses previous values of iterates to compute predicted updates.

Tangent Initialization Methods

SimpleContinuation.BorderedInitialTangent — Type

BorderedInitialTangent{I} <: AbstractInitialTangent

Constructs initial tangent using the bordered method, using every base-vector.

Fields

positive_search_order::Bool: Whether to search for a base-vector in the positive or negative direction. Default: true

search_order::I: A user provided vector of integers specifying the order in which to search for a base-vector. If not provided, the default is to search in the positive direction.

SimpleContinuation.BorderedInitialTangent — Method

BorderedInitialTangent(search_order::I) where {I<:AbstractVector}

Constructor for BorderedInitialTangent. Sets positive_search_order to true. See struct docs for details.

Arguments

search_order::I: A user provided vector of integers specifying the order in which to search for a base-vector. Must be an integer vector.

SimpleContinuation.BorderedInitialTangent — Method

BorderedInitialTangent(; positive_search_order=true)

Constructor for BorderedInitialTangent. Sets search_order to nothing. See struct docs for details.

Kwargs

positive_search_order::Bool: Whether to search for a base-vector in the positive or negative direction. Default: true

SimpleContinuation.SecantInitialTangent — Type

SecantInitialTangent <: AbstractInitialTangent

Constructs initial tangent using the secant method with a single natural continuation step.

Fields

step_scale_factor::Float64: The scale factor used to perturb the natural continuation parameter, as a fraction of the initial step size.

SimpleContinuation.SecantInitialTangent — Method

SecantInitialTangent(; step_scale_factor=1e-6)

Constructor for SecantInitialTangent.

Kwargs

step_scale_factor::Float64: The scale factor used to perturb the natural continuation parameter, as a fraction of the initial step size. Default: 1e-6

SimpleContinuation.UserInitialTangent — Type

UserInitialTangent <: AbstractInitialTangent

Constructs initial tangent using a user provided tangent vector.

SimpleContinuation.UserInitialTangent — Method

UserInitialTangent(user_initial_tangent::AbstractVector)

Constructor for UserInitialTangent. See struct docs for details.

Arguments

user_initial_tangent::AbstractVector: A user provided tangent vector. Must be a vector of length n + 1, where n is the dimension of the system.

Step Limiters

Methods to limit the length of the correction step with respect to the size of the tangent. Helps to avoid erroneous curve switching.

SimpleContinuation.CorrectionStepLimiter — Type

struct CorrectionStepLimiter

A type to limit the step size in the correction step.

Restricts the Euclidean distance between the prediction and correction to be less than some frac of Δs at the current iteration. If the distance is larger, the correction step is rejected despite success of the Newton-Raphson solver.

Fields

frac::Float64: The fraction of the current PALC step size Δs that the correction step length should be less than.

SimpleContinuation.CorrectionStepLimiter — Method

function CorrectionStepLimiter(; frac=0.1)

Constructor for the correction step limiter.

Kwargs

frac::Float64: The fraction of the current PALC step size Δs that the correction step length should be less than. Defaults to 0.1.

Inner Products

Types of inner products.

SimpleContinuation.BifurcationKitInnerProduct — Type

BifurcationKitInnerProduct <: AbstractInnerProduct

A scaled Euclidean inner product used in BifurcationKit.jl.

Fields

θ::Float64: The scaling factor for the inner product. Must be in [0,1].

SimpleContinuation.DoubleScaledInnerProduct — Type

DoubleScaledInnerProduct <: AbstractInnerProduct

A double scaled Euclidean inner product. The scaling factor θ scales the relative contributions of the unknowns and the continuation parameter to the inner product.

Fields

θ::Float64: The scaling factor for the inner product. Must be in [0,1].

SimpleContinuation.ScaledInnerProduct — Type

ScaledInnerProduct <: AbstractInnerProduct

A scaled Euclidean inner product. The scaling factor θ scales the relative contributions of the unknowns and the continuation parameter to the inner product.

Fields

θ::Float64: The scaling factor for the inner product. Must be in [0,1].

SimpleContinuation.StandardDotProduct — Type

StandardDotProduct <: AbstractInnerProduct

A standard, unscaled Euclidean inner product. This is the default inner product used in PALC.

Trace

Options for REPL printing during continuation runs

SimpleContinuation.Silent — Type

Silent <: AbstractTraceLevel

No output is printed to the console during continuation. This is the default for PALC.

SimpleContinuation.ContinuationSteps — Type

ContinuationSteps <: AbstractTraceLevel

Prints basic information at the end of each step.

SimpleContinuation.ContinuationAndNewtonSteps — Type

ContinuationAndNewtonSteps <: AbstractTraceLevel

Prints the same information as ContinuationSteps, but also prints information about the Newton iterations at each step.